pyproject.toml
[project]
name = "regscan"
version = "0.1.0"
description = "Regulatory Intelligence Engine - 메인화면 콘텐츠/피드용 정보 생성"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dateutil>=2.8.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9",
    "pandas>=1.5.0",
    "apscheduler>=3.10.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-playwright>=0.4.0",
    "ruff>=0.1.0",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]
crawl = [
    "playwright>=1.40.0",
    "parsel>=1.8.0",
]
gcs = [
    "google-cloud-storage>=2.14.0",
]
gemini = [
    "google-generativeai>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
